Desktop/GitHub-Projetos/HypersDAO/requirements.txt
python-telegram-bot==21.7
requests==2.31.0
schedule==1.2.0
httpx==0.27.0
